+++ fake-cucumber/javascript/src/FakeTestResultsStream.ts
@@ -10,7 +10,7 @@ class FakeTestResultsStream extends Transform {
       | 'ndjson'
       | 'protobuf'
       | 'protobuf-objects',
-    private readonly results: 'none' | 'random' | 'pattern',
+    private readonly results: 'none' | 'random' | 'pattern'
   ) {
     super({ objectMode: true })
   }
@@ -18,7 +18,7 @@ class FakeTestResultsStream extends Transform {
   public _transform(
     envelope: messages.IEnvelope,
     encoding: string,
-    callback: (error?: Error | null, data?: any) => void,
+    callback: (error?: Error | null, data?: any) => void
   ): void {
     this.p(envelope)
 
@@ -28,7 +28,7 @@ class FakeTestResultsStream extends Transform {
           testCaseStarted: new messages.TestCaseStarted({
             pickleId: envelope.pickle.id,
           }),
-        }),
+        })
       )
       let index = 0
       let testCaseStatus: messages.TestResult.Status =
@@ -43,7 +43,7 @@ class FakeTestResultsStream extends Transform {
               index,
               pickleId: envelope.pickle.id,
             }),
-          }),
+          })
         )
 
         switch (this.results) {
@@ -59,6 +59,68 @@ class FakeTestResultsStream extends Transform {
           default:
             throw new Error(`Unexpected results value: ${this.results}`)
         }
+
+        if (
+          ![
+            messages.TestResult.Status.UNDEFINED,
+            messages.TestResult.Status.AMBIGUOUS,
+            messages.TestResult.Status.UNKNOWN,
+          ].includes(testStepStatus)
+        ) {
+          const stepMatchArguments: messages.IStepMatchArgument[] = []
+
+          let word = ''
+          let wordIndex = 0
+          step.text.split('').forEach((character, i) => {
+            if (character !== ' ') {
+              word += character
+            } else {
+              if (word.length > 0) {
+                if (wordIndex % 2 === 1) {
+                  stepMatchArguments.push(
+                    new messages.StepMatchArgument({
+                      group: new messages.StepMatchArgument.Group({
+                        value: word,
+                        start: i - word.length,
+                      }),
+                      parameterTypeName: 'fake',
+                    })
+                  )
+                }
+                word = ''
+                wordIndex++
+              }
+            }
+          })
+
+          if (word.length > 0) {
+            stepMatchArguments.push(
+              new messages.StepMatchArgument({
+                group: new messages.StepMatchArgument.Group({
+                  value: word,
+                  start: step.text.length - word.length,
+                }),
+                parameterTypeName: 'fake',
+              })
+            )
+            word = ''
+          }
+
+          this.p(
+            new messages.Envelope({
+              testStepMatched: new messages.TestStepMatched({
+                index,
+                pickleId: envelope.pickle.id,
+                stepDefinitionReference: new messages.SourceReference({
+                  location: new messages.Location({ column: 4, line: 999 }),
+                  uri: envelope.pickle.uri + '.cobol',
+                }),
+                stepMatchArguments,
+              }),
+            })
+          )
+        }
+
         if (testStepStatus > testCaseStatus) {
           testCaseStatus = testStepStatus
         }
@@ -70,10 +132,14 @@ class FakeTestResultsStream extends Transform {
               pickleId: envelope.pickle.id,
               testResult: {
                 status: testStepStatus,
-                message: testStepStatus === messages.TestResult.Status.FAILED ? `Some error message\n\tfake_file:2\n\tfake_file:7\n` : null,
+                message:
+                  testStepStatus === messages.TestResult.Status.FAILED
+                    ? `Some error message\n\tfake_file:2\n\tfake_file:7\n`
+                    : null,
+                durationNanoseconds: 123456789,
               },
             }),
-          }),
+          })
         )
         index++
       }
@@ -84,10 +150,14 @@ class FakeTestResultsStream extends Transform {
             pickleId: envelope.pickle.id,
             testResult: {
               status: testCaseStatus,
-              message: testStepStatus === messages.TestResult.Status.FAILED ? `Some error message\n\tfake_file:2\n\tfake_file:7\n` : null,
+              message:
+                testStepStatus === messages.TestResult.Status.FAILED
+                  ? `Some error message\n\tfake_file:2\n\tfake_file:7\n`
+                  : null,
+              durationNanoseconds: 987654321,
             },
           }),
-        }),
+        })
       )
     }