+++ cucumber-messages/CHANGELOG.md
@@ -9,7 +9,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+* New `TestStepMatched` message
+
 ### Changed
+
+* Renamed `PatternMatch` to `StepMatchArgument`
+* Renamed `CommandRunTestStep#patternMatches` to `CommandRunTestStep#stepMatchArguments`
 
 ### Deprecated