+++ gherkin/CHANGELOG.md
@@ -13,11 +13,16 @@ This document is formatted according to the principles of [Keep A CHANGELOG](htt
 
 ### Changed
 
+* Upgrade cucumber-messages to 6.0.0
+
 ### Deprecated
 
 ### Removed
 
 ### Fixed
+
+* [Go] Fix trimming of whitespace in example cells
+  ([#703](https://github.com/cucumber/cucumber/pull/703) [aslakhellesoy])
 
 ## [7.0.4] - 2019-08-29